requirement.txt
Django==5.2.1
python-dotenv==1.1.0
psycopg2-binary==2.9.10
djangorestframework==3.16.0

